Fix PSR-4 violations in sources
