+++ test/ZipStreamerTest.php
@@ -8,7 +8,8 @@
  */
 namespace ZipStreamer;
 
-require_once  __DIR__ . "/../src/ZipStreamer.php";
+use ZipStreamer\ZipStreamer;
+
 require_once  __DIR__ . "/ZipComponents.php";
 
 class File {
@@ -31,7 +32,7 @@ class File {
   }
 }
 
-class TestZipStreamer extends \PHPUnit_Framework_TestCase {
+class TestZipStreamer extends \PHPUnit\Framework\TestCase {
   const ATTR_MADE_BY_VERSION = 0x032d; // made by version (upper byte: UNIX, lower byte v4.5)
   const EXT_FILE_ATTR_DIR = 0x41ed0010;
   const EXT_FILE_ATTR_FILE = 0x81a40000;
